move all tests concerning Axis class out of grid.py
